refactor filepath
